Partial changes for repository implementation

+++ data/src/main/kotlin/local/converters/GenreIdsConverter.kt
@@ -5,12 +5,12 @@ import androidx.room.TypeConverter
 class GenreIdsConverter {
 
     @TypeConverter
-    fun fromString(value: String): List<Int> {
-        return value.split(",").map { it.toInt() }
+    fun fromString(value: String): List<Long> {
+        return value.split(",").map { it.toLong() }
     }
 
     @TypeConverter
-    fun fromList(value: List<Int>): String {
+    fun fromList(value: List<Long>): String {
         return value.joinToString(",")
     }
 }
